requirements.txt
oyaml
PyGithub
requests
Flask
python-jenkins
whaaaaat
